worker/requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
httpx>=0.24.0
openai>=1.0.0
